Fix error messages on migrated fields
The current/old macros all displayed the first
error in the field.errors list.

The original implementation of the migrated fields
displayed all of them as a single string. This is
a change in behaviour and broke one of the tests
for a radios field.

This changes it back, to match the original
behaviour.