✨ Add /blocklist support to notify API client
To help avoid abuse of the CAST Notify service, as well as protect our
accounts with backend providers like `firetext`, we want to apply some
restrictions on the Sender IDs that users can apply to our outbound SMS
and emails.

To support the above functionality, a new `/services/
{service-id}/blocklist has been added to the notifications API, allowing
retrieval of a list containing all blocked phrases, for example `GOVUK`
or `NHS`.

This commit updates our `ServiceAPIClient` model, adding support for the
new blocklist API endpoint.

https://app.asana.com/0/1199707043388412/1199940470983800