✨ Introduce new SenderBlocklistValidator class
To help avoid abuse of the CAST Notify service, as well as protect our
accounts with backend providers like `firetext`, we want to apply some
restrictions on the Sender IDs that users can apply to our outbound SMS
and emails.

This commit introduces a new `WTForms` `Validator` class, allowing the
user's submitted Sender Id to be screened against a blocklist. If the
requested Sender ID matches against any of the blocked phrases,
validation will fail and the user will be shown a meaningful error.

https://app.asana.com/0/1199707043388412/1199940470983800