🐛 Fix Flake8 Linting Failures
In a recent PR, we implemented a "blocklist" feature where the
`notifications-api` returns a blocklist of "protected" terms,
referenced by the `notifications-admin` app to enforce restrictions on
SMS sender values added by the user. Unfortunately, the new code for
this feature has triggered a handful of Flake8 linting errors which must
be resolved to allow successful passing of the test-suite.

This commit addresses the following errors from the Flake8 linting
library:

1) Imported libraries must be ordered alphabetically.
2) Class definitions must be separated by 2 empty lines.
3) Maximum line length must not be longer than 120 characters.
4) Using mutable values are default constructor arguments.

https://app.asana.com/0/1199707043388412/1199940470983783