✨ Install NPM version manager in bootstrap.sh
The `notifications-admin` application uses NPM (Node Package Manager) to
manage front-end dependencies. It's possible the developer's operating
system already has npm installed, though it may not be the version we
desire. To solve this problem, we can use the "n" tool to manage
multiple node versions. In our case, we set the node to run the latest
lts (long term support) release.

This commit moves the installation and use of "n" within the
`bootstrap.sh` script, ensuring the command gets run and hiding needless
complexity from the user (developer).